Projects/polymarket_arb_bot_v2/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
py-clob-client>=0.25.0
websocket-client>=1.8.0
